Correct linter Errors
