+++ json/books.json
@@ -1,17 +1,7 @@
 [
   {
-    "publisher": "Abdi",
-    "cover_state": "good",
-    "publish_date": "2020-03-09"
-  },
-  {
-    "publisher": "Adams",
-    "cover_state": "best",
-    "publish_date": "2020-09-02"
-  },
-  {
-    "publisher": "Pacifix",
-    "cover_state": "better",
-    "publish_date": "2020-08-08"
+    "publisher": "ethos",
+    "cover_state": "old",
+    "publish_date": "2000-09-09"
   }
 ]
