+++ json/labels.json
@@ -1,14 +1,6 @@
 [
   {
-    "title": "smile",
-    "color": "red"
-  },
-  {
-    "title": "purpose",
-    "color": "pacific"
-  },
-  {
-    "title": "don't give up",
+    "title": "Hope",
     "color": "red"
   }
 ]
