+++ main.rb
@@ -1,5 +1,6 @@
 require_relative 'app'
 require_relative 'options'
+
 def main
   app = App.new
   app.load
