+++ options.rb
@@ -1,11 +1,12 @@
+require_relative 'app'
+
 class Options
   def choices
     puts 'WELCOME TO CATALOG OF MY THINGS APP'
     puts 'Choose an option by entering its number'
-    puts '1. List all games         2. List all books           3. List all music albums'
-    puts '4. List all movies        5. List all genres          6. List all labels '
-    puts '7. List all authors       8. List all sources         9. Add a book'
-    puts '10. Add an author         11. Add a music album       12. Add a movie'
-    puts '13. Add a game            14.Add a Label              [0] Quit'
+    puts '1. List all games         2. List all books            3. List all music albums'
+    puts '4. List all genres        5. List all labels           6. List all authors '
+    puts '7. Add a book             8. Add a music album         9. Add a game'
+    puts '0. Quit'
   end
 end
